feat: update currency handling
Updated the way how we handle the currency in the system. Now instead of directly getting the
user's currency or the default one, we use the getRuntimeCurrency() method. This method is likely
to return the correct currency even if the user did not set it or if there is no default currency
set in the system. This change is applied in several files including `index.php`,
`CurrencySwitch.php`, and `LangCurrencySwitch.php`. Additionally, updated `composer.json` to use
version 2.2 of `quiqqer/currency`. Removed unused code.